+++ .env-example
@@ -11,9 +11,9 @@ DW_USERNAME = [USER]
 DW_PASSWORD = [PASS]
 
 # Prefect flows
-PREFECT_LOGGING_EXTRA_LOGGERS = "orderline_dw_scripts"
+PREFECT_LOGGING_EXTRA_LOGGERS = "orderline_dw"
 PREFECT_API_URL = "http://127.0.0.1:4545/api"
 
 # Prefect flows op Docker
-PREFECT_LOGGING_EXTRA_LOGGERS = "orderline_dw_scripts"
+PREFECT_LOGGING_EXTRA_LOGGERS = "orderline_dw"
 PREFECT_API_URL = "http://host.docker.internal:4545"
