+++ dbt_project.yml
@@ -19,7 +19,8 @@ clean-targets:         # directories to be removed by `dbt clean`
   - "target"
   - "dbt_packages"
 
-
+vars:
+  sales_materialization: 'incremental'
 
 models:
   orderline_dw_dbt:
