+++ docker-compose.yml
@@ -1,0 +1,34 @@
+version: '3.8'
+
+services:
+  postgresdb:
+    image: postgres:latest
+    environment:
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_DB: prefect
+    volumes:
+      - postgres-data:/var/lib/postgresql/data
+    networks:
+      - orderline-dw-network
+
+  prefectserver:
+    build:
+      context: .
+      dockerfile: docker/prefectserver/Dockerfile
+    depends_on:
+      - postgresdb
+    environment:
+      PREFECT_SERVER_API_HOST: 0.0.0.0
+      # PREFECT_SERVER_API_PORT: 4444
+      PREFECT_API_DATABASE_CONNECTION_URL: "postgresql+asyncpg://postgres:postgres@postgresdb:5432/prefect"
+    ports:
+      - "4200:4200"
+    networks:
+      - orderline-dw-network
+
+networks:
+  orderline-dw-network:
+
+volumes:
+  postgres-data:
