+++ orderline_dw/dbt/models/dm_sales/sales.sql
@@ -1,3 +1,9 @@
+{{ config(materialized=var('sales_materialization')) }}
+
 SELECT
     *
 FROM {{ ref('sales_t1') }}
+{% if is_incremental() %}
+    WHERE bestelmoment > (SELECT MAX(bestelmoment) FROM {{ this }})
+{% endif %}
+
