+++ orderline_dw/my_prefect/all.py
@@ -1,12 +1,12 @@
 from prefect import flow
 
-from orderline_dw.my_prefect import initial_load
+from orderline_dw.my_prefect import load
 from orderline_dw.my_prefect import init
 
 @flow
 def all_flow():
     init.init_flow()
-    initial_load.initial_load_flow()
+    load.full_load_flow()
 
 
 if __name__ == "__main__":
